backend/requirements.txt
fastapi
uvicorn
openai
aiohttp
python-dotenv
requests
supabase

